+++ packages/ic_siws_js/src/context.type.ts
@@ -1,14 +1,21 @@
 import { DelegationChain, DelegationIdentity } from "@dfinity/identity";
-import type { LoginStatus, PrepareLoginStatus } from "./state.type";
+import type { Adapter } from "@solana/wallet-adapter-base";
+import type { PublicKey } from "@solana/web3.js";
 
-import { PublicKey } from "@solana/web3.js";
+export type PrepareLoginStatus = "error" | "preparing" | "success" | "idle";
+export type LoginStatus = "error" | "logging-in" | "success" | "idle";
+export type SignMessageStatus = "error" | "idle" | "pending" | "success";
 
 export type SiwsIdentityContextType = {
   /** Is set to `true` on mount until a stored identity is loaded from local storage or
    * none is found. */
   isInitializing: boolean;
 
-  /** Load a SIWS message from the provider canister, to be used for login. Calling prepareLogin
+  /** Sets the wallet adapter to be used for signing the SIWE message. This must be called
+   * before calling `login()`. Alternatively, you can pass the adapter to the SiwsManager constructor. */
+  setAdapter: (adapter: Adapter) => Promise<void>;
+
+  /** Load a SIWE message from the provider canister, to be used for login. Calling prepareLogin
    * is optional, as it will be called automatically on login if not called manually. */
   prepareLogin: () => void;
 
@@ -30,7 +37,7 @@ export type SiwsIdentityContextType = {
   /** Error that occurred during the prepareLogin process. */
   prepareLoginError?: Error;
 
-  /** Initiates the login process by requesting a SIWS message from the backend. */
+  /** Initiates the login process by requesting a SIWE message from the backend. */
   login: () => Promise<DelegationIdentity | undefined>;
 
   /** Reflects the current status of the login process. */
@@ -51,6 +58,12 @@ export type SiwsIdentityContextType = {
   /** Error that occurred during the login process. */
   loginError?: Error;
 
+  /** Status of the SIWE message signing process. */
+  signMessageStatus: SignMessageStatus;
+
+  /** Error that occurred during the SIWE message signing process. */
+  signMessageError?: Error;
+
   /** The delegation chain is available after successfully loading the identity from local
    * storage or completing the login process. */
   delegationChain?: DelegationChain;
@@ -62,7 +75,7 @@ export type SiwsIdentityContextType = {
   /** The Ethereum address associated with current identity. This address is not necessarily
    * the same as the address of the currently connected wallet - on wallet change, the addresses
    * will differ. */
-  identityAddress?: PublicKey;
+  identityPublicKey?: PublicKey;
 
   /** Clears the identity from the state and local storage. Effectively "logs the user out". */
   clear: () => void;