+++ packages/ic_siws_js/src/service.interface.ts
@@ -13,12 +13,24 @@ export interface Delegation {
 
 export type GetDelegationResponse = { Ok: SignedDelegation } | { Err: string };
 
-export interface LoginDetails {
-  user_canister_pubkey: CanisterPublicKey;
+export interface LoginOkResponse {
   expiration: Timestamp;
+  user_canister_pubkey: CanisterPublicKey;
 }
 
-export type LoginResponse = { Ok: LoginDetails } | { Err: string };
+export type LoginResponse = { Ok: LoginOkResponse } | { Err: string };
+
+export interface SiwsMessage {
+  domain: string;
+  address: Address;
+  statement: string;
+  uri: string;
+  version: number;
+  chain_id: string;
+  nonce: string;
+  issued_at: bigint;
+  expiration_time: bigint;
+}
 
 export type PrepareLoginResponse = { Ok: SiwsMessage } | { Err: string };
 
@@ -30,26 +42,21 @@ export interface SignedDelegation {
   signature: Uint8Array | number[];
   delegation: Delegation;
 }
-export interface SiwsMessage {
-  uri: string;
-  issued_at: bigint;
-  domain: string;
-  statement: string;
-  version: number;
-  chain_id: string;
-  address: Address;
-  nonce: string;
-  expiration_time: bigint;
-}
 
 export type SiwsSignature = string;
 
 export type Timestamp = bigint;
+
+export type Nonce = string;
+
 export interface SIWS_IDENTITY_SERVICE {
+  siws_prepare_login: ActorMethod<[Address], PrepareLoginResponse>;
+  siws_login: ActorMethod<
+    [SiwsSignature, Address, SessionKey, Nonce],
+    LoginResponse
+  >;
   siws_get_delegation: ActorMethod<
     [Address, SessionKey, Timestamp],
     GetDelegationResponse
   >;
-  siws_login: ActorMethod<[SiwsSignature, Address, SessionKey], LoginResponse>;
-  siws_prepare_login: ActorMethod<[Address], PrepareLoginResponse>;
 }