+++ packages/ic_siws_provider/src/service/siws_login.rs
@@ -16,17 +16,19 @@ use crate::{update_root_hash, ADDRESS_PRINCIPAL, PRINCIPAL_ADDRESS, SETTINGS, ST
 ///
 /// # Arguments
 /// * `signature` (String): The signature of the SIWS message.
-/// * `address` (String): The Solana address of the user.
+/// * `pubkey` (String): The Solana pubkey of the user.
 /// * `session_key` (ByteBuf): A unique key that identifies the session.
+/// * `nonce` (String): The nonce generated during the `prepare_login` call.
 ///
 /// # Returns
-/// * `Ok(LoginOkResponse)`: Contains the user canister public key and other login response data if the login is successful.
+/// * `Ok(LoginDetails)`: Contains the session expiration time and user canister public key if the login is successful.
 /// * `Err(String)`: An error message if the login process fails.
 #[update]
 fn siws_login(
     signature: String,
     pubkey: String,
     session_key: ByteBuf,
+    nonce: String,
 ) -> Result<LoginDetails, String> {
     STATE.with(|state| {
         let signature_map = &mut *state.signature_map.borrow_mut();
@@ -43,6 +45,7 @@ fn siws_login(
             session_key,
             &mut *signature_map,
             &ic_cdk::api::id(),
+            &nonce,
         )
         .map_err(|e| e.to_string())?;